+++ stroom-core-server/src/main/java/stroom/entity/server/util/ConnectionUtil.java
@@ -18,6 +18,7 @@ package stroom.entity.server.util;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Supplier;
+import org.springframework.jdbc.datasource.DataSourceUtils;
 import stroom.entity.shared.BaseResultList;
 import stroom.entity.shared.HasPrimitiveValue;
 import stroom.entity.shared.PrimitiveValueConverter;
@@ -29,6 +30,7 @@ import stroom.util.logging.StroomLogger;
 import stroom.util.shared.ModelStringUtil;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.DriverManager;
@@ -39,6 +41,7 @@ import java.sql.Statement;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
@@ -81,7 +84,17 @@ public class ConnectionUtil {
         return DriverManager.getConnection(driverUrl, driverUsername, driverPassword);
     }
 
-    public static final int getMultiInsertMaxBatchSize() {
+    /**
+     * Attempts to get the connection currently in use by a transaction using the passed datasource, or create one
+     * if there isn't one.
+     * @param dataSource The dataSource to get the connection to
+     * @return A new or existing connection
+     */
+    public static Connection getConnection(final DataSource dataSource) {
+        return DataSourceUtils.getConnection(dataSource);
+    }
+
+    private static int getMultiInsertMaxBatchSize() {
         return StroomProperties.getIntProperty(MULTI_INSERT_BATCH_SIZE, MULTI_INSERT_BATCH_SIZE_DEFAULT);
     }
 
@@ -115,7 +128,7 @@ public class ConnectionUtil {
 
             preparedStatement.close();
 
-            log(logExecutionTime, result, () -> sql, args);
+            log(logExecutionTime, () -> Integer.toString(result), () -> sql, args);
 
             return result;
         } catch (final SQLException sqlException) {
@@ -146,6 +159,43 @@ public class ConnectionUtil {
         } catch (final SQLException sqlException) {
             LOGGER.error("executeUpdate() - " + sql + " " + args, sqlException);
             throw sqlException;
+        }
+    }
+
+    public static void executeStatement(final Connection connection, final String sql) throws SQLException {
+        executeStatements(connection, Collections.singletonList(sql));
+    }
+
+    public static void executeStatements(final Connection connection, final List<String> sqlStatements) throws SQLException {
+        LOGGER.debug(">>> %s", sqlStatements);
+        final LogExecutionTime logExecutionTime = new LogExecutionTime();
+        try (Statement statement = connection.createStatement()) {
+
+            sqlStatements.forEach(sql -> {
+                try {
+                    statement.addBatch(sql);
+                } catch (SQLException e) {
+                    throw new RuntimeException(String.format("Error adding sql [%s] to batch", sql), e);
+                }
+            });
+            int[] results = statement.executeBatch();
+            boolean isFailure = Arrays.stream(results)
+                    .anyMatch(val -> val == Statement.EXECUTE_FAILED);
+
+            if (isFailure) {
+                throw new RuntimeException(String.format("Got error code for batch %s", sqlStatements));
+            }
+
+            log(logExecutionTime,
+                    () -> Arrays.stream(results)
+                            .mapToObj(Integer::toString)
+                            .collect(Collectors.joining(",")),
+                    sqlStatements::toString,
+                    Collections.emptyList());
+
+        } catch (final Exception e) {
+            LOGGER.error("executeStatement() - " + sqlStatements, e);
+            throw e;
         }
     }
 
@@ -221,7 +271,7 @@ public class ConnectionUtil {
             PreparedStatementUtil.setArguments(preparedStatement, args);
             final ResultSet resultSet = PreparedStatementUtil.createCloseStatementResultSet(preparedStatement);
 
-            log(logExecutionTime, null, sql, args);
+            log(logExecutionTime, "ResultSet", sql, args);
 
             return resultSet;
         } catch (final SQLException sqlException) {
@@ -234,17 +284,17 @@ public class ConnectionUtil {
                             final Object result,
                             final String sql,
                             final List<Object> args) {
-        log(logExecutionTime, result, () -> sql, args);
+        log(logExecutionTime, result::toString, () -> sql, args);
 
     }
     private static void log(final LogExecutionTime logExecutionTime,
-                            final Object result,
+                            final Supplier<String> resultSupplier,
                             final Supplier<String> sqlSupplier,
                             final List<Object> args) {
         final long time = logExecutionTime.getDuration();
         if (LOGGER.isDebugEnabled() || time > 1000) {
             final String message = "<<< " + sqlSupplier.get() + " " + args + " took " + ModelStringUtil.formatDurationString(time)
-                    + " with result " + result;
+                    + " with result " + resultSupplier.get();
             if (time > 1000) {
                 LOGGER.warn(message);
             } else {
@@ -431,7 +481,7 @@ public class ConnectionUtil {
                     keyList = Collections.emptyList();
                 }
 
-                log(logExecutionTime, result, preparedStatement::toString, allArgs);
+                log(logExecutionTime, () -> Integer.toString(result), preparedStatement::toString, allArgs);
                 return keyList;
 
             } catch (final SQLException sqlException) {