+++ stroom-integrationtest/src/test/java/stroom/DatabaseCommonTestControl.java
@@ -36,6 +36,7 @@ import stroom.index.server.IndexShardWriterCache;
 import stroom.index.shared.Index;
 import stroom.index.shared.IndexShard;
 import stroom.index.shared.IndexShardService;
+import stroom.jobsystem.shared.ClusterLock;
 import stroom.jobsystem.shared.Job;
 import stroom.jobsystem.shared.JobNode;
 import stroom.lifecycle.LifecycleServiceImpl;
@@ -50,11 +51,13 @@ import stroom.pipeline.shared.PipelineEntity;
 import stroom.pipeline.shared.TextConverter;
 import stroom.pipeline.shared.XSLT;
 import stroom.script.shared.Script;
+import stroom.security.server.AppPermission;
 import stroom.security.server.DocumentPermission;
 import stroom.security.server.Permission;
 import stroom.security.server.User;
 import stroom.security.server.UserGroupUser;
-import stroom.statistics.shared.StatisticStore;
+import stroom.statistics.shared.StatisticStoreEntity;
+import stroom.streamstore.server.StreamAttributeValueFlush;
 import stroom.streamstore.server.fs.FileSystemUtil;
 import stroom.streamstore.shared.FindStreamAttributeKeyCriteria;
 import stroom.streamstore.shared.Stream;
@@ -77,6 +80,7 @@ import javax.annotation.Resource;
 import java.io.File;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
@@ -90,6 +94,44 @@ import java.util.Map;
 public class DatabaseCommonTestControl implements CommonTestControl, ApplicationContextAware {
     private static final StroomLogger LOGGER = StroomLogger.getLogger(DatabaseCommonTestControl.class);
 
+    private static final List<String> TABLES_TO_CLEAR = Arrays.asList(
+            AppPermission.TABLE_NAME,
+            ClusterLock.TABLE_NAME,
+            Dashboard.TABLE_NAME,
+            Dictionary.TABLE_NAME,
+            DocumentPermission.TABLE_NAME,
+            Feed.TABLE_NAME,
+            Folder.TABLE_NAME,
+            Index.TABLE_NAME,
+            Index.TABLE_NAME_INDEX_VOLUME, //link table between IDX and VOL so no entity of its own
+            IndexShard.TABLE_NAME,
+            Job.TABLE_NAME,
+            JobNode.TABLE_NAME,
+            Node.TABLE_NAME,
+            Permission.TABLE_NAME,
+            PipelineEntity.TABLE_NAME,
+            Query.TABLE_NAME,
+            Rack.TABLE_NAME,
+            Res.TABLE_NAME,
+            Script.TABLE_NAME,
+            StatisticStoreEntity.TABLE_NAME,
+            Stream.TABLE_NAME,
+            StreamAttributeKey.TABLE_NAME,
+            StreamAttributeValue.TABLE_NAME,
+            StreamProcessor.TABLE_NAME,
+            StreamProcessorFilter.TABLE_NAME,
+            StreamProcessorFilterTracker.TABLE_NAME,
+            StreamTask.TABLE_NAME,
+            StreamVolume.TABLE_NAME,
+            TextConverter.TABLE_NAME,
+            User.TABLE_NAME,
+            UserGroupUser.TABLE_NAME,
+            Visualisation.TABLE_NAME,
+            Volume.TABLE_NAME,
+            VolumeState.TABLE_NAME,
+            XMLSchema.TABLE_NAME,
+            XSLT.TABLE_NAME);
+
     @Resource
     private VolumeService volumeService;
     @Resource
@@ -98,6 +140,8 @@ public class DatabaseCommonTestControl implements CommonTestControl, Application
     private ImportExportSerializer importExportSerializer;
     @Resource
     private StreamAttributeKeyService streamAttributeKeyService;
+    @Resource
+    private StreamAttributeValueFlush streamAttributeValueFlush;
     @Resource
     private IndexShardManager indexShardManager;
     @Resource
@@ -123,6 +167,8 @@ public class DatabaseCommonTestControl implements CommonTestControl, Application
     @Override
     public void setup() {
         Instant startTime = Instant.now();
+        //ensure the constraints are enabled in case teardown did not happen on a previous test
+        databaseCommonTestControlTransactionHelper.enableConstraints();
         nodeConfig.setup();
         createStreamAttributeKeys();
 
@@ -137,54 +183,12 @@ public class DatabaseCommonTestControl implements CommonTestControl, Application
     @Override
     public void teardown() {
         Instant startTime = Instant.now();
-        deleteEntity(StreamTask.class);
-
-        deleteEntity(StreamVolume.class);
-        deleteEntity(StreamAttributeValue.class);
-        deleteEntity(Stream.class);
-
-        deleteEntity(Query.class);
-        deleteEntity(Dashboard.class);
-        deleteEntity(Visualisation.class);
-        deleteEntity(Script.class);
-        deleteEntity(Res.class);
-        deleteEntity(Dictionary.class);
-        deleteEntity(StatisticStore.class);
-
         // Make sure we are no longer creating tasks.
         streamTaskCreator.shutdown();
 
         // Make sure we don't delete database entries without clearing the pool.
         indexShardWriterCache.shutdown();
         indexShardManager.deleteFromDisk();
-
-        deleteEntity(IndexShard.class);
-        deleteEntity(Index.class);
-
-        deleteEntity(Feed.class);
-
-        deleteEntity(XMLSchema.class);
-        deleteEntity(TextConverter.class);
-        deleteEntity(XSLT.class);
-
-        deleteEntity(StreamProcessorFilter.class);
-        deleteEntity(StreamProcessorFilterTracker.class);
-        deleteEntity(StreamProcessor.class);
-
-        deleteEntity(PipelineEntity.class);
-
-        deleteEntity(UserGroupUser.class);
-        deleteEntity(DocumentPermission.class);
-        deleteEntity(Permission.class);
-        deleteEntity(User.class);
-
-        // Delete folders last as they are the parent for many other entities.
-        deleteEntity(Folder.class);
-
-        // deleteTable("sys_user_role");
-        // deleteTable("sys_user_group");
-        // deleteTable("sys_user");
-        // deleteTable("sys_group");
 
         // Delete the contents of all volumes.
         final List<Volume> volumes = volumeService.find(new FindVolumeCriteria());
@@ -194,16 +198,18 @@ public class DatabaseCommonTestControl implements CommonTestControl, Application
             FileSystemUtil.deleteContents(FileSystemUtil.createFileTypeRoot(volume).getParentFile());
         }
 
-        // These are static
-        deleteEntity(JobNode.class);
-        deleteEntity(Job.class);
+        //Clear any un-flushed stream attributes
+        streamAttributeValueFlush.clear();
 
-        deleteEntity(Volume.class);
-        deleteEntity(VolumeState.class);
-        deleteEntity(Node.class);
-        deleteEntity(Rack.class);
-
+        //ensure any hibernate entities are flushed down before we clear the tables
         databaseCommonTestControlTransactionHelper.clearContext();
+
+        //clear all the tables using direct sql on a different connection
+        //in theory trncating the tables should be quicker but it was takeing 1.5s to trancate all the tables
+        //so used delete with no constraint checks instead
+        databaseCommonTestControlTransactionHelper.clearTables(TABLES_TO_CLEAR);
+
+        //ensure all the caches are empty
         stroomCacheManager.clear();
 
         final Map<String, Clearable> clearableBeanMap = applicationContext.getBeansOfType(Clearable.class, false,
@@ -217,6 +223,8 @@ public class DatabaseCommonTestControl implements CommonTestControl, Application
     public void createStreamAttributeKeys() {
         final BaseResultList<StreamAttributeKey> list = streamAttributeKeyService
                 .find(new FindStreamAttributeKeyCriteria());
+        LOGGER.info("Existing stream attribute keys: [%s]", list);
+
         final HashSet<String> existingItems = new HashSet<String>();
         for (final StreamAttributeKey streamAttributeKey : list) {
             existingItems.add(streamAttributeKey.getName());