+++ stroom-integrationtest/src/test/java/stroom/TestAbstractCoreIntegrationTest.java
@@ -1,0 +1,12 @@
+package stroom;
+
+import org.junit.Test;
+
+public class TestAbstractCoreIntegrationTest extends AbstractCoreIntegrationTest {
+
+
+    @Test
+    public void testSetupAndTearDown() throws Exception {
+        //do nothing, just checking to make sure setup and teardown work in isolation
+    }
+}